sync

+++ commander.cpp
@@ -7,12 +7,14 @@
 #include "commander.h"
 #include "settings.h"
 #include "localfile.h"
+#include "version.h"
 
 using namespace Sara;
 
 Commander::Commander(QObject *parent)
     : QObject(parent)
 {
+    m_bCopy = false;
     m_pProcess = new QProcess(this);
     connect(m_pProcess, SIGNAL(readyReadStandardError()), this, SIGNAL(processError()));
     connect(m_pProcess, SIGNAL(readyReadStandardOutput()), this, SIGNAL(processOutput()));
@@ -48,6 +50,7 @@ bool Commander::run(const Sara::Update& aUpdate)
     QString command;
     QStringList commandParameters;
 
+    m_bCopy = false;
     m_oUpdate = aUpdate;
 
     command = setCommandBasedOnOS();
@@ -74,14 +77,49 @@ bool Commander::run(const Sara::Update& aUpdate)
     qDebug() << "command: " << command;
     qDebug() << "commandlline: " << commandParameters.join(" ");
 
-    m_pProcess->start(command, commandParameters);
+    commandParameters.removeAll("");
 
-    // wait 5 minutes for process start
-    if(!m_pProcess->waitForStarted(1000 * 60 * 5))
+    if(m_bCopy && commandParameters.size() > 1)
     {
-        emit progressText(tr("Error: Update '%1' failed to start").arg(m_oUpdate.getTitle()));
-        m_pProcess->kill();
-        return false;
+        QFile file(commandParameters.at(0));
+        QFile destFile(commandParameters.at(1));
+
+        if(!file.exists())
+        {
+            emit updateExit(-1, QProcess::NormalExit);
+            return false;
+        }
+
+        if(destFile.exists())
+            if(!destFile.remove())
+            {
+                emit updateExit(-2, QProcess::NormalExit);
+                return false;
+            }
+
+        if(file.copy(commandParameters.at(1)))
+        {
+            emit updateExit(0, QProcess::NormalExit);
+            return true;
+        }
+        else
+        {
+            qDebug() << file.errorString();
+            emit updateExit(file.error(), QProcess::NormalExit);
+            return false;
+        }
+    }
+    else
+    {
+        m_pProcess->start(command, commandParameters);
+
+        // wait 5 minutes for process start
+        if(!m_pProcess->waitForStarted(1000 * 60 * 5))
+        {
+            emit progressText(tr("Error: Update '%1' failed to start").arg(m_oUpdate.getTitle()));
+            m_pProcess->kill();
+            return false;
+        }
     }
     return true;
 }
@@ -97,7 +135,7 @@ QString Commander::readStdOut() const
 }
 
 /// command = "ld.exe [HOME] [INI@/home/user/.config/Sara/Client.conf:uuid] [SHELL] [@[HOME]/.config/Sara/Client.conf:uuid]";
-QString Commander::resolve(const QString& aString) const
+QString Commander::resolve(const QString& aString)
 {
     QString theString = aString;
 
@@ -109,7 +147,11 @@ QString Commander::resolve(const QString& aString) const
     theString = theString.replace("[SARA_UP_TARGETCODE]", m_oUpdate.getTargetVersion().getCode());
     theString = theString.replace("[SARA_UP_TYPE]", QString::number(m_oUpdate.getType()));
     theString = theString.replace("[SARA_FILE]", Sara::LocalFile::getDownloadLocation(m_oUpdate.getDownloadLink()));
+    theString = theString.replace("[SARA_FILENAME]", QFileInfo(Sara::LocalFile::getDownloadLocation(m_oUpdate.getDownloadLink())).fileName());
+    theString = theString.replace("[SARA_FILEEXT]", QFileInfo(Sara::LocalFile::getDownloadLocation(m_oUpdate.getDownloadLink())).completeSuffix());
+    theString = theString.replace("[SARA_SEP]", QDir::separator());
     theString = theString.replace("[SARA_PATH]", Sara::LocalFile::getDownloadPath());
+    theString = theString.replace("[SARA_VERSION]", SARA_CLIENT_VERSION);
 
     // replace environment vars
     QStringList env = QProcessEnvironment::systemEnvironment().toStringList();
@@ -154,6 +196,11 @@ QString Commander::resolve(const QString& aString) const
         pos = theString.indexOf(QString("[@"), pos+1);
     }
 
+    if(theString.indexOf("[SARA_COPY_COMMAND]")>-1)
+        m_bCopy = true;
+
+    // remove letfovers
+    theString = theString.replace(QRegExp("\\[[^]]*\\]"), "");
+
     return theString;
 }
-
